+++ fuzztest/internal/runtime.cc
@@ -1083,7 +1083,7 @@ bool FuzzTestFuzzerImpl::RunInUnitTestMode(const Configuration& configuration) {
     const auto time_limit = stats_.start_time + duration;
     Input mutation{params_domain_.Init(prng)};
     const size_t max_iterations = duration == absl::ZeroDuration() ? 0 : 10000;
-    for (int i = 0; i < max_iterations; ++i) {
+    for (int i = 0; i < (int)max_iterations; ++i) {
       runtime_.SetExternalFailureDetected(false);
       RunOneInput(mutation);
       if (runtime_.external_failure_detected()) {
